refactor: Moved source code files to src/ subfolder
Moved ragbot.py, ragbot_streamlit.py, and helpers.py to the new src/ subfolder. The script paths for running ragbot.py and ragbot_streamlit.py have been updated to include the "src" directory. This ensures that the correct script is executed when running the shell shortcuts for RagBot and the web browser version. I did this to better organize the code and also in preparation for adding unit tests. Having the source code files in a separate folder also helps use ragbot better on itself.